+++ app/Http/Controllers/Admin/PlatformController.php
@@ -6,21 +6,36 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\Platform\StorePlatformRequest;
 use App\Http\Requests\Admin\Platform\UpdatePlatformRequest;
 use App\Models\Platform;
+use App\Services\FileUploadService;
 use Exception;
 use Illuminate\Contracts\View\Factory;
 use Illuminate\Contracts\View\View;
 use Illuminate\Foundation\Application;
 use Illuminate\Http\RedirectResponse;
+use Illuminate\Http\Request;
 use Illuminate\Support\Facades\DB;
 
 class PlatformController extends Controller
 {
+    protected mixed $uploadPath;
+    public function __construct(private readonly FileUploadService $fileUpload)
+    {
+        $this->uploadPath = config('upload.platform_path');
+        $this->middleware('permission:platforms-index', ['only' => ['index']]);
+        $this->middleware('permission:platforms-create', ['only' => ['store']]);
+        $this->middleware('permission:platforms-edit', ['only' => ['update']]);
+        $this->middleware('permission:platforms-delete', ['only' => ['destroy']]);
+    }
     /**
      * Display a listing of the resource.
      */
-    public function index(): View|Application|Factory
+    public function index(Request $request): View|Application|Factory
     {
-        $platforms = Platform::latest()->paginate(10);
+        $query = Platform::query();
+        if ($request->input('q')) {
+            $query->search('title', trim(request()->input('q')));
+        }
+        $platforms = $query->latest()->paginate(15)->withQueryString();
 
         return view('admin.platforms.index', compact('platforms'));
     }
@@ -33,29 +48,25 @@ class PlatformController extends Controller
         try {
             DB::beginTransaction();
 
-            $imageFileName = null;
-
-            if ($request->has('image')) {
-                $imageName = generateFileName($request->image->getClientOriginalName());
-                $request->file('image')->storeAs(env('CATEGORY_IMAGE_PATH'), $imageName, 'public');
+            $imageName = null;
+            if ($request->hasFile('image')) {
+                $imageName = $this->fileUpload->upload($request->file('image'), $this->uploadPath);
             }
 
             Platform::create([
-                'title' => $request->input('title'),
-                'is_active' => $request->input('is_active'),
-                'image' => $imageFileName,
+                ...$request->validated(),
+                'image' => $imageName
             ]);
 
             DB::commit();
-        } catch (Exception $ex) {
+        } catch (Exception $e) {
             DB::rollBack();
-            toastr()->error($ex->getMessage().'مشکلی پیش آمد!');
-
-            return redirect()->back();
+            report($e);
+            flash()->error(config('flasher.platform.create_failed'));
+            return redirect()->back()->withInput();
         }
 
-        toastr()->success('با موفقیت اضافه شد!');
-
+        toastr()->success(config('flasher.platform.created'));
         return redirect()->back();
     }
 
@@ -67,29 +78,24 @@ class PlatformController extends Controller
         try {
             DB::beginTransaction();
 
-            if ($request->has('image')) {
-                $imageName = generateFileName($request->image->getClientOriginalName());
-                $request->file('image')->storeAs(env('CATEGORY_IMAGE_PATH'), $imageName, 'public');
+            if ($request->hasFile('image')) {
+                $imageName = $this->fileUpload->replace($request->file('image'), $platform->image, $this->uploadPath);
                 $platform->update([
                     'image' => $imageName,
                 ]);
             }
 
-            $platform->update([
-                'title' => $request->title,
-                'is_active' => $request->is_active,
-            ]);
+            $platform->update($request->validated());
 
             DB::commit();
-        } catch (Exception $ex) {
+        } catch (Exception $e) {
             DB::rollBack();
-            toastr()->error('مشکلی پیش آمد!', $ex->getMessage());
-
-            return redirect()->back();
+            report($e);
+            flash()->error(config('flasher.platform.update_failed'));
+            return redirect()->back()->withInput();
         }
 
-        toastr()->success('با موفقیت ویرایش شد.');
-
+        flash()->success(config('flasher.platform.updated'));
         return redirect()->back();
     }
 
@@ -98,10 +104,24 @@ class PlatformController extends Controller
      */
     public function destroy(Platform $platform): RedirectResponse
     {
-        $platform->delete();
+        if ($platform->products()->count() > 0) {
+            flash()->success('پلتفرم قابل حذف نیسش');
+            return redirect()->back();
+        }else{
+            try {
+                $this->fileUpload->delete(
+                    $this->uploadPath,
+                    $platform->image
+                );
+                $platform->delete();
 
-        toastr()->success('با موفقیت حذف شد!');
-
-        return redirect()->back();
+                flash()->success(config('flasher.platform.deleted'));
+                return redirect()->back();
+            } catch (Exception $ex) {
+                report($ex);
+                flash()->error(config('flasher.platform.delete_failed'));
+                return redirect()->back();
+            }
+        }
     }
 }
